Disable digital noise gate (fixes CYAN-4506).
This fixes CYAN-4505 (Sound cuts out at low volume) [1].
One might think that the digital noise gate (DNG) is better left enabled to
reduce the static noise but in my tests, only the "payload" audio was cut off,
while the static continued. For me it seems that this feature of the sound chip
was not intended for use in a smartphone.

[1]: https://jira.cyanogenmod.org/browse/CYAN-4506.

Change-Id: I4e41c594695f85574ed35e2aef59d6b768403606