galaxys2-common: Add prebuilt uncompressed ramdisk and recovery
We need to separate builds for the embedded ramdisk.cpio
and recovery-ramdisk-device.cpio in zImage. When building the
kernel with 'make bootimage' the ramdisk.cpio and
recovery-ramdisk-device.cpio must already be built. In Q we
can't realize a dependency for this. Therefore we must create
prebuilt images before we built the ROM/kernel.

Use the following commands to update these images:
croot
make ramdisk-galaxys2-common
make recoveryimage-galaxys2-common

Change-Id: I56e364d6e0a03d145673f5656d8a61659e01d452